backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
paramiko>=3.4.0
psutil>=5.9.0
pynvml>=11.5.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0
